Long overdue rehaul of presentation compiler tests. Added support for options file, and switched simple-tests to use the STARR library on the test classpath.
Put any options you need passed to the presentation compiler in a file
with the same name as the test, ending in '.opts'. Currently it's used
to set the -bootclasspath for simple-tests. Check the contents of
simple-tests.opts for all the options.

IMPORTANT: when committing a new STARR, you need to check this test. Most
likely, you *will* break it otherwise. 

review by extempore

git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@24360 5e8d7ff9-d8ef-0310-90f0-a4852d11357a