Rooting out mismatched zips.
I added local logging to zip and zipped and listened for who was
dropping things on the floor. Everything in this commit stems from that.
Sometimes the fix was uncertain and I sprinkled some logging. If you've
been hanging back with lots of internals knowledge waiting for the right
commit to review, this would be a good one. But since knowledgeable
people are hard to find, I'll go with review by moors.

git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25688 5e8d7ff9-d8ef-0310-90f0-a4852d11357a