Fixes #4351.
Added an "Abstract" method info to the specialized phase, which denotes that no implementation should be generated.

Previously:

trait A[@specialized(Boolean) T] {
  def foo: T
}

used to generate:

trait A$mcZ$sp extends A[Boolean] {
  def foo$mcZ$sp = this.foo
}

which caused cyclic calls because of linearization rules when several traits are mixed together. Now, the following is generated:

trait A$mcZ$sp extends A[Boolean] {
  def foo$mcZ$sp: Boolean
}

Review by dragos.